+++ src/pii_extract/defs.py
@@ -8,3 +8,6 @@ COUNTRY_ANY = "any"
 FMT_CONFIG_PLUGIN = "pii-extract:plugins:v1"
 FMT_CONFIG_TASKS = "pii-extract:tasks:v1"
 FMT_CONFIG_TASKCFG = "pii-extract:task-config:v1"
+
+# Stage name for PII entity "process" field
+STAGE = "detection"